Fix TestFilesystem::read_dir_diff_paths
This function collects the files present in the `left` and `right` directories.
It should return the path suffix with the `left` and `right` prefix stripped.
This is necessary because `process_opts` will add the `left` and `right` prefix
back to the paths returned by `read_dir_diff_paths`. If we don't strip the
prefix, we will end up with invalid paths where the prefix is duplicated.

This went unnoticed since the current tests don't diff directories. I will add
a test that diffs directories in a subsequent commit.