ui: clear screen at start of session
When using `scm-record` over Mosh, the screen does not get cleared. I
don't know why it's not always a problem.

I don't know much about `ratatui`, but I noticed on
https://docs.rs/ratatui/latest/ratatui/widgets/struct.Clear.html that
it assumes that the drawing area is clear and that `Terminal::clear()`
can be used to clear it otherwise, so that's what this PR does.