feat: Add bare-minimum help dialog
Pressing question mark pops up a dialog that tells you you can use the
mouse to click menu items to do things and view mouse shortcuts.

Obviously this is not perfect, but it would have saved me a bunch of
frustration, and it does a fair amount of the grunt work towards a
better help box.

One thing that made me stop here is that the main event loop should
probably be split up to do different things depending on if any given
dialog is open. Adding the HelpDialog using the current paradigm would
have involved approximately tripling the size of the match for no
benefit.