Add ctrl-left and ctrl-h as keybindings to move outwards
These are currently duplicative of the unqualified versions of left and h. In
the next commit however, we will make left and h collapse the current text
section if it is expanded, or move outwards if it is already collapsed. Adding
the ctrl keybindings will give users a way to jump out without collapsing the
section.

For consistency, ctrl-right and ctrl-l are also supported as keybindings to
move the focus inwards.