Add a small table of key bindings to the "?" help dialog
Extended `Dialog` to accept a `Text`, and used that to create a styled
table of key bindings with manual layout.

Ideally, `Dialog` would take a widget and wrap it in a centered block, but
I couldn't figure out how to do that with ratatui, and I think this is
still a good first (second) step.

Table from: https://paulsmith.github.io/jj-builtin-diff-editor-cheatsheet/